refactor; change graph serialization method
