+++ scratchpad/pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-python_files = contr_sheme.py expr_json.py test_*.py
+python_files = contr_sheme.py expr_*.py test_*.py
